Crate evm_domain_test_runtime

Modules

api

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

BaseFeeThreshold

BlockGasLimit

EVM block gas limit is set to maximum to allow all the transaction stored on Consensus chain.

BlockHashCount

BoundDivision

ChannelInitReservePortion

ChannelReserveFee

CheckExtrinsicsValidityError

DecodeExtrinsicError

DefaultBaseFeePerGas

DefaultElasticity

DomainChainByteFee

EVMCurrencyAdapter

ExistentialDeposit

ExtrinsicStorageFees

FinalDomainTransactionByteFee

FindAuthorTruncated

MaxLocks

MaxReserves

MmrProofVerifier

OnXDMRewards

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PostOnlyBlockHash

PrecompilesValue

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SelfChainId

SessionKeys

StorageKeys

SuicideQuickClearLimit

TransactionConverter

TransporterEndpointId

Version

WeightPerGas

Enums

HoldIdentifier

Balance hold identifier for this runtime.

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

GAS_PER_SECOND

Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU).

RUNTIME_API_VERSIONS

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_GAS

Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Functions

extract_signer

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balance of an account.

Balances

BalancesConfig

BaseFee

BaseFeeConfig

Block

Block type as expected by this runtime.

BlockFees

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

EVM

EVMChainId

EVMChainIdConfig

EVMConfig

EVMNoncetracker

Ethereum

EthereumConfig

Executive

Executive: handles dispatch to the various modules.

ExecutivePallet

Hash

A hash of some data used by the chain.

Header

Opaque block header type.

Messenger

Nonce

Index of a transaction in the chain.

Precompiles

Precompiles we use for EVM

SelfDomainId

SelfDomainIdConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Transporter

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.