Crate evm_domain_test_runtime

Modules

api

opaque

Opaque types.

Structs

BlockHashCount

ChannelInitReservePortion

ChannelReserveFee

CheckExtrinsicsValidityError

DecodeExtrinsicError

DomainChainByteFee

EVMCurrencyAdapter

ExistentialDeposit

ExtrinsicStorageFees

FinalDomainTransactionByteFee

FindAuthorTruncated

HoldIdentifierWrapper

Balance hold identifier for this runtime.

IntoRuntimeCall

MaxLocks

MaxOutgoingMessages

MaxReserves

MinimumTransfer

MmrProofVerifier

OnXDMRewards

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PostOnlyBlockHash

PrecompilesValue

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SelfChainId

SessionKeys

StorageKeys

TransactionConverter

TransactionWeightFee

TransporterEndpointId

UnbalancedHandler

UnbalancedHandler that will just burn any unbalanced funds

Version

Enums

HoldIdentifier

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

RUNTIME_API_VERSIONS

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Functions

construct_extrinsic_raw_payload

extract_signer

Type Aliases

AccountId

Some way of identifying an account on the EVM chain. We intentionally make it equivalent to the public key of the EVM transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balance

Balance of an account.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockFees

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

EVM

EVMChainId

EVMChainIdConfig

EVMConfig

EVMNoncetracker

Ethereum

EthereumConfig

EvmGasPriceCalculator

Executive

Executive: handles dispatch to the various modules.

ExecutivePallet

Hash

A hash of some data used by the chain.

Header

Opaque block header type.

Messenger

Nonce

Index of a transaction in the chain.

Precompiles

Precompiles we use for EVM

SelfDomainId

SelfDomainIdConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the EVM chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

StorageOverlayChecks

Sudo

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Transporter

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility