Crate evm_domain_test_runtime

Modules

api

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

AccountId20Converter

BaseFeeThreshold

BlockGasLimit

EVM gas limit

BlockHashCount

BoundDivision

CoreDomainId

DefaultBaseFeePerGas

DefaultElasticity

ExistentialDeposit

FindAuthorTruncated

MaxLocks

MaxReserves

MaximumRelayers

OperationalFeeMultiplier

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PostOnlyBlockHash

PrecompilesValue

RelayConfirmationDepth

RelayerDeposit

Runtime

RuntimeApi

RuntimeApiImpl

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SessionKeys

StateRootsBound

TransactionByteFee

TransactionConverter

TransporterEndpointId

Version

WeightPerGas

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

Constants

EXISTENTIAL_DEPOSIT

The existential deposit. Same with the one on primary chain.

GAS_PER_SECOND

Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU). Given the 500ms Weight, from which 75% only are used for transactions, the total EVM execution gas limit is: GAS_PER_SECOND * 0.500 * 0.75 ~= 15_000_000.

MAXIMUM_BLOCK_LENGTH

MAXIMUM_BLOCK_WEIGHT

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_MILLISECS_PER_BLOCK

We allow for 2000ms of compute with a 6 second average block time.

WEIGHT_PER_GAS

Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Functions

extract_signer

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Balance

Balance of an account.

Balances

BalancesConfig

BaseFee

BaseFeeConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BlockNumber

An index to a block.

CheckedExtrinsic

Extrinsic type that has already been checked.

EVM

EVMChainId

EVMChainIdConfig

EVMConfig

Ethereum

EthereumConfig

Executive

Executive: handles dispatch to the various modules.

ExecutivePallet

GenesisConfig

Hash

A hash of some data used by the chain.

Header

Opaque block header type.

Index

Index of a transaction in the chain.

Messenger

MessengerConfig

Precompiles

Precompiles we use for EVM

SelfDomainId

SelfDomainIdConfig

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Transporter

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.